+++ packages/eyes-sdk-core/index.js
@@ -2,8 +2,76 @@
 
 /* eslint-disable max-len */
 
-const common = require('@applitools/eyes-common')
+// config
+exports.AccessibilityLevel = require('./lib/config/AccessibilityLevel').AccessibilityLevel
+exports.AccessibilityGuidelinesVersion = require('./lib/config/AccessibilityGuidelinesVersion').AccessibilityGuidelinesVersion
+exports.AccessibilityMatchSettings = require('./lib/config/AccessibilityMatchSettings').AccessibilityMatchSettings
+exports.AccessibilityRegionType = require('./lib/config/AccessibilityRegionType').AccessibilityRegionType
+exports.BatchInfo = require('./lib/config/BatchInfo').BatchInfo
+exports.BrowserType = require('./lib/config/BrowserType').BrowserType
+exports.Configuration = require('./lib/config/Configuration').Configuration
+exports.DeviceName = require('./lib/config/DeviceName').DeviceName
+exports.ExactMatchSettings = require('./lib/config/ExactMatchSettings').ExactMatchSettings
+exports.FloatingMatchSettings = require('./lib/config/FloatingMatchSettings').FloatingMatchSettings
+exports.ImageMatchSettings = require('./lib/config/ImageMatchSettings').ImageMatchSettings
+exports.MatchLevel = require('./lib/config/MatchLevel').MatchLevel
+exports.PropertyData = require('./lib/config/PropertyData').PropertyData
+exports.ProxySettings = require('./lib/config/ProxySettings').ProxySettings
+exports.ScreenOrientation = require('./lib/config/ScreenOrientation').ScreenOrientation
+exports.SessionType = require('./lib/config/SessionType').SessionType
+exports.StitchMode = require('./lib/config/StitchMode').StitchMode
+exports.IosDeviceName = require('./lib/config/IosDeviceName').IosDeviceName
+exports.IosScreenOrientation = require('./lib/config/IosScreenOrientation').IosScreenOrientation
+exports.IosVersion = require('./lib/config/IosVersion').IosVersion
 
+// debug
+exports.DebugScreenshotsProvider = require('./lib/debug/DebugScreenshotsProvider').DebugScreenshotsProvider
+exports.FileDebugScreenshotsProvider = require('./lib/debug/FileDebugScreenshotsProvider').FileDebugScreenshotsProvider
+exports.NullDebugScreenshotProvider = require('./lib/debug/NullDebugScreenshotProvider').NullDebugScreenshotProvider
+
+// errors
+exports.EyesError = require('./lib/errors/EyesError').EyesError
+
+// geometry
+exports.CoordinatesType = require('./lib/geometry/CoordinatesType').CoordinatesType
+exports.Location = require('./lib/geometry/Location').Location
+exports.RectangleSize = require('./lib/geometry/RectangleSize').RectangleSize
+exports.Region = require('./lib/geometry/Region').Region
+
+// handler
+exports.PropertyHandler = require('./lib/handler/PropertyHandler').PropertyHandler
+exports.ReadOnlyPropertyHandler = require('./lib/handler/ReadOnlyPropertyHandler').ReadOnlyPropertyHandler
+exports.SimplePropertyHandler = require('./lib/handler/SimplePropertyHandler').SimplePropertyHandler
+
+// images
+exports.ImageDeltaCompressor = require('./lib/images/ImageDeltaCompressor').ImageDeltaCompressor
+exports.MutableImage = require('./lib/images/MutableImage').MutableImage
+
+// logging
+exports.ConsoleLogHandler = require('./lib/logging/ConsoleLogHandler').ConsoleLogHandler
+exports.DebugLogHandler = require('./lib/logging/DebugLogHandler').DebugLogHandler
+exports.FileLogHandler = require('./lib/logging/FileLogHandler').FileLogHandler // -browser
+exports.Logger = require('./lib/logging/Logger').Logger
+exports.LogHandler = require('./lib/logging/LogHandler').LogHandler
+exports.NullLogHandler = require('./lib/logging/NullLogHandler').NullLogHandler
+
+// useragent
+exports.BrowserNames = require('./lib/useragent/BrowserNames').BrowserNames
+exports.OSNames = require('./lib/useragent/OSNames').OSNames
+exports.UserAgent = require('./lib/useragent/UserAgent').UserAgent
+
+// utils
+exports.ArgumentGuard = require('./lib/utils/ArgumentGuard').ArgumentGuard
+exports.ConfigUtils = require('./lib/utils/ConfigUtils').ConfigUtils
+exports.DateTimeUtils = require('./lib/utils/DateTimeUtils').DateTimeUtils
+exports.FileUtils = require('./lib/utils/FileUtils').FileUtils
+exports.GeneralUtils = require('./lib/utils/GeneralUtils').GeneralUtils
+exports.ImageUtils = require('./lib/utils/ImageUtils').ImageUtils
+exports.PerformanceUtils = require('./lib/utils/PerformanceUtils').PerformanceUtils
+exports.StreamUtils = require('./lib/utils/StreamUtils')
+exports.TypeUtils = require('./lib/utils/TypeUtils').TypeUtils
+exports.deserializeDomSnapshotResult = require('./lib/utils/deserializeDomSnapshotResult')
+exports.DomCapture = require('./lib/DomCapture').DomCapture
 exports.AppOutputProvider = require('./lib/capture/AppOutputProvider').AppOutputProvider
 exports.AppOutputWithScreenshot = require('./lib/capture/AppOutputWithScreenshot').AppOutputWithScreenshot
 exports.EyesScreenshot = require('./lib/capture/EyesScreenshot').EyesScreenshot
@@ -33,10 +101,10 @@ exports.NewTestError = require('./lib/errors/NewTestError').NewTestError
 exports.OutOfBoundsError = require('./lib/errors/OutOfBoundsError').OutOfBoundsError
 exports.TestFailedError = require('./lib/errors/TestFailedError').TestFailedError
 exports.EyesDriverOperationError = require('./lib/errors/EyesDriverOperationError').EyesDriverOperationError
+exports.ElementNotFoundError = require('./lib/errors/ElementNotFoundError').ElementNotFoundError
 
 exports.CheckSettings = require('./lib/fluent/CheckSettings')
-exports.CheckSettingsFactory = require('./lib/fluent/CheckSettingsFactory')
-exports.FrameLocator = require('./lib/fluent/FrameLocator')
+exports.DriverCheckSettings = require('./lib/fluent/DriverCheckSettings')
 exports.locatorToPersistedRegions = require('./lib/fluent/locatorToPersistedRegions')
 exports.GetRegion = require('./lib/fluent/GetRegion').GetRegion
 exports.GetSelector = require('./lib/fluent/GetSelector').GetSelector
@@ -51,7 +119,6 @@ exports.GetAccessibilityRegion = require('./lib/fluent/GetAccessibilityRegion').
 exports.AccessibilityRegionByRectangle = require('./lib/fluent/AccessibilityRegionByRectangle').AccessibilityRegionByRectangle
 exports.AccessibilityRegionBySelector = require('./lib/fluent/AccessibilityRegionBySelector')
 exports.AccessibilityRegionByElement = require('./lib/fluent/AccessibilityRegionByElement')
-exports.TargetRegionBySelector = require('./lib/fluent/TargetRegionBySelector')
 exports.TargetRegionByElement = require('./lib/fluent/TargetRegionByElement')
 
 exports.AppOutput = require('./lib/match/AppOutput').AppOutput
@@ -78,8 +145,6 @@ exports.CssTranslatePositionProvider = require('./lib/positioning/CssTranslatePo
 exports.ScrollElementPositionProvider = require('./lib/positioning/ScrollElementPositionProvider')
 exports.CssTranslateElementPositionProvider = require('./lib/positioning/CssTranslateElementPositionProvider')
 exports.PositionMemento = require('./lib/positioning/PositionMemento')
-exports.ScrollPositionMemento = require('./lib/positioning/ScrollPositionMemento')
-exports.CssTranslatePositionMemento = require('./lib/positioning/CssTranslatePositionMemento')
 
 exports.RenderInfo = require('./lib/renderer/RenderInfo').RenderInfo
 exports.RenderRequest = require('./lib/renderer/RenderRequest').RenderRequest
@@ -111,91 +176,31 @@ exports.Trigger = require('./lib/triggers/Trigger').Trigger
 
 exports.AppEnvironment = require('./lib/AppEnvironment').AppEnvironment
 exports.EyesBase = require('./lib/EyesBase').EyesBase
+exports.EyesClassic = require('./lib/EyesClassic')
+exports.EyesVisualGrid = require('./lib/EyesVisualGrid')
+exports.EyesFactory = require('./lib/EyesFactory')
 exports.EyesJsBrowserUtils = require('./lib/EyesJsBrowserUtils').EyesJsBrowserUtils
 exports.EyesUtils = require('./lib/EyesUtils')
 exports.FailureReports = require('./lib/FailureReports').FailureReports
 exports.MatchSingleWindowTask = require('./lib/MatchSingleWindowTask').MatchSingleWindowTask
 exports.MatchWindowTask = require('./lib/MatchWindowTask').MatchWindowTask
-exports.RenderWindowTask = require('./lib/RenderWindowTask').RenderWindowTask
 exports.TestResults = require('./lib/TestResults').TestResults
-exports.TestAccessibilityStatus = require('./lib/TestResults').TestAccessibilityStatus
+exports.TestResultsError = require('./lib/TestResults').TestResultsError
 exports.AccessibilityStatus = require('./lib/AccessibilityStatus').AccessibilityStatus
 exports.TestResultsFormatter = require('./lib/TestResultsFormatter').TestResultsFormatter
 exports.TestResultsStatus = require('./lib/TestResultsStatus').TestResultsStatus
 
-exports.FrameChain = require('./lib/frames/FrameChain').FrameChain
-exports.Frame = require('./lib/frames/Frame').Frame
+exports.FrameChain = require('./lib/frames/FrameChain')
+exports.Frame = require('./lib/frames/Frame')
 
-exports.EyesWrappedDriver = require('./lib/wrappers/EyesWrappedDriver').EyesWrappedDriver
-exports.EyesWrappedElement = require('./lib/wrappers/EyesWrappedElement').EyesWrappedElement
-exports.EyesJsExecutor = require('./lib/wrappers/EyesJsExecutor').EyesJsExecutor
-exports.EyesElementFinder = require('./lib/wrappers/EyesElementFinder').EyesElementFinder
-exports.EyesBrowsingContext = require('./lib/wrappers/EyesBrowsingContext').EyesBrowsingContext
+exports.EyesWrappedDriver = require('./lib/wrappers/EyesWrappedDriver')
+exports.EyesWrappedElement = require('./lib/wrappers/EyesWrappedElement')
+exports.EyesJsExecutor = require('./lib/wrappers/EyesJsExecutor')
+exports.EyesElementFinder = require('./lib/wrappers/EyesElementFinder')
+exports.EyesBrowsingContext = require('./lib/wrappers/EyesBrowsingContext')
 
 exports.EyesRunner = require('./lib/runner/EyesRunner').EyesRunner
 exports.ClassicRunner = require('./lib/runner/ClassicRunner').ClassicRunner
 exports.VisualGridRunner = require('./lib/runner/VisualGridRunner').VisualGridRunner
 exports.TestResultContainer = require('./lib/runner/TestResultContainer').TestResultContainer
 exports.TestResultsSummary = require('./lib/runner/TestResultsSummary').TestResultsSummary
-
-// Classes from eyes-common which may be used as part of public API
-exports.AccessibilityLevel = common.AccessibilityLevel
-exports.AccessibilityMatchSettings = common.AccessibilityMatchSettings
-exports.AccessibilityRegionType = common.AccessibilityRegionType
-exports.BatchInfo = common.BatchInfo
-exports.BrowserType = common.BrowserType
-exports.Configuration = common.Configuration
-exports.DeviceName = common.DeviceName
-exports.ExactMatchSettings = common.ExactMatchSettings
-exports.FloatingMatchSettings = common.FloatingMatchSettings
-exports.ImageMatchSettings = common.ImageMatchSettings
-exports.MatchLevel = common.MatchLevel
-exports.PropertyData = common.PropertyData
-exports.ProxySettings = common.ProxySettings
-exports.ScreenOrientation = common.ScreenOrientation
-exports.SessionType = common.SessionType
-exports.StitchMode = common.StitchMode
-exports.IosVersion = common.IosVersion
-exports.IosScreenOrientation = common.IosScreenOrientation
-exports.IosDeviceName = common.IosDeviceName
-
-exports.DebugScreenshotsProvider = common.DebugScreenshotsProvider
-exports.FileDebugScreenshotsProvider = common.FileDebugScreenshotsProvider
-exports.NullDebugScreenshotProvider = common.NullDebugScreenshotProvider
-
-exports.EyesError = common.EyesError
-
-exports.CoordinatesType = common.CoordinatesType
-exports.Location = common.Location
-exports.RectangleSize = common.RectangleSize
-exports.Region = common.Region
-
-exports.PropertyHandler = common.PropertyHandler
-exports.ReadOnlyPropertyHandler = common.ReadOnlyPropertyHandler
-exports.SimplePropertyHandler = common.SimplePropertyHandler
-
-exports.ImageDeltaCompressor = common.ImageDeltaCompressor
-exports.MutableImage = common.MutableImage
-
-exports.ConsoleLogHandler = common.ConsoleLogHandler
-exports.DebugLogHandler = common.DebugLogHandler
-exports.FileLogHandler = common.FileLogHandler
-exports.Logger = common.Logger
-exports.LogHandler = common.LogHandler
-exports.NullLogHandler = common.NullLogHandler
-
-// Classes which can be used internally, but should not be exported from final SDKs
-exports.BrowserNames = common.BrowserNames
-exports.OSNames = common.OSNames
-exports.UserAgent = common.UserAgent
-
-exports.ArgumentGuard = common.ArgumentGuard
-exports.ConfigUtils = common.ConfigUtils
-exports.DateTimeUtils = common.DateTimeUtils
-exports.FileUtils = common.FileUtils
-exports.GeneralUtils = common.GeneralUtils
-exports.ImageUtils = common.ImageUtils
-exports.PerformanceUtils = common.PerformanceUtils
-exports.StreamUtils = common.StreamUtils
-exports.TypeUtils = common.TypeUtils
-exports.deserializeDomSnapshotResult = common.deserializeDomSnapshotResult