+++ packages/eyes-sdk-core/lib/capture/SafariScreenshotImageProvider.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const {Location, Region, OSNames} = require('@applitools/eyes-common')
+const {Location, Region, OSNames} = require('../..')
 const {ImageProvider} = require('./ImageProvider')
 const EyesUtils = require('../EyesUtils')
 
@@ -78,7 +78,7 @@ class SafariScreenshotImageProvider extends ImageProvider {
       const frameChain = this._driver.context.frameChain
       let loc =
         frameChain.size > 0
-          ? frameChain.getTopFrameScrollLocation()
+          ? frameChain.first.parentScrollLocation
           : await EyesUtils.getScrollLocation(this._logger, this._driver.executor).catch(
               () => Location.ZERO,
             )