+++ packages/eyes-sdk-core/test/unit/server/SessionStartInfo.spec.js
@@ -14,17 +14,15 @@ const {
   RectangleSize,
   FloatingMatchSettings,
   AccessibilityLevel,
+  AccessibilityGuidelinesVersion,
   AccessibilityRegionType,
   AccessibilityRegionByRectangle,
-  CheckSettingsFactory,
+  DriverCheckSettings,
   MatchWindowTask,
 } = require('../../../index')
 const {EyesBaseImpl} = require('../../testUtils')
 
-const CheckSettings = CheckSettingsFactory(
-  function() {},
-  function() {},
-)
+const CheckSettings = DriverCheckSettings.specialize({WrappedElement: () => {}, Frame: () => {}})
 
 describe('SessionStartInfo', () => {
   it('TestSerialization', () => {
@@ -71,7 +69,10 @@ describe('SessionStartInfo', () => {
             maxRightOffset: 20,
           }),
         ],
-        accessibilityLevel: AccessibilityLevel.AA,
+        accessibilitySettings: {
+          level: AccessibilityLevel.AA,
+          guidelinesVersion: AccessibilityGuidelinesVersion.WCAG_2_0,
+        },
       }),
       branchName: 'some branch',
       parentBranchName: 'parent branch',
@@ -85,7 +86,7 @@ describe('SessionStartInfo', () => {
       properties,
     })
 
-    const actualSerialization = JSON.stringify(sessionStartInfo)
+    const actualSerialization = JSON.stringify(sessionStartInfo, null, 2)
     const expectedSerialization = getResourceAsText('SessionStartInfo_Serialization.json')
     assert.strictEqual(
       actualSerialization,