+++ packages/eyes-cypress/CHANGELOG.md
@@ -2,12 +2,34 @@
 
 ## Unreleased
 
+## 3.12.0 - 2020/6/9
+
+- Unified core
+- updated to @applitools/dom-snapshot@3.5.3 (from 3.5.2)
+- updated to @applitools/eyes-sdk-core@11.0.5 (from 10.3.1)
+- updated to @applitools/visual-grid-client@14.4.4 (from v14.2.1)
+
+## 3.11.1 - 2020/5/26
+
+- updated to @applitools/eyes-sdk-core@10.3.1 (from v10.2.0)
+- updated to @applitools/visual-grid-client@14.2.1 (from v14.1.0)
+
+## 3.11.0 - 2020/5/19
+
+- Support for accessibility validation
+- updated to @applitools/eyes-sdk-core@10.2.0
+- updated to @applitools/visual-grid-client@14.1.0
+
+ ## 3.10.5 - 2020/5/18
+
+- fixed typescript definitions file for eyesCheckWindow string argument and open browser name parameter. 
+- updated to @applitools/eyes-sdk-core@10.1.2
+- updated to @applitools/visual-grid-client@14.0.2
 
 ## 3.10.4 - 2020/5/7
 
 - updated to @applitools/dom-snapshot@3.5.2
 - updated to @applitools/eyes-sdk-core@10.0.1
-- updated to @applitools/visual-grid-client@13.7.6
 - updated to @applitools/visual-grid-client@13.7.7
 
  ## 3.10.3 - 2020/4/30