+++ packages/eyes-images/CHANGELOG.md
@@ -3,6 +3,18 @@
 ## Unreleased
 
 
+
+## 4.14.0 - 2020/6/9
+
+- updated to @applitools/eyes-sdk-core@11.0.2 (from v10.2.0)
+- updated to @applitools/eyes-sdk-core@11.0.4 (from v11.0.2)
+- updated to @applitools/eyes-sdk-core@11.0.5 (from 11.0.4)
+
+## 4.13.0 - 2020/5/19
+
+- Support for AccessibilityGuidelinesVersion
+- updated to @applitools/eyes-sdk-core@10.2.0 from 9.0.2
+
 ## 4.12.38
 
 - upload domsnapshot directly to Azure [Trello](https://trello.com/c/ZCLJo8Fy/241-upload-dom-directly-to-azure)