+++ packages/eyes-selenium/CHANGELOG.md
@@ -4,6 +4,34 @@
 
 
 
+## 4.39.0 - 2020/6/9
+
+- added mobile web API support for VG
+- updated to @applitools/eyes-sdk-core@11.0.4
+- updated to @applitools/visual-grid-client@14.4.3
+- updated to @applitools/eyes-sdk-core@11.0.5 (from 11.0.4)
+- updated to @applitools/visual-grid-client@14.4.4 (from 14.4.3)
+
+## 4.38.0 - 2020/6/2
+
+- Unified core
+- updated to @applitools/eyes-sdk-core@11.0.2 (from v10.3.0)
+- updated to @applitools/visual-grid-client@14.4.1 (from v14.2.0)
+
+## 4.37.0 - 2020/5/19
+
+- Support for AccessibilityGuidelinesVersion
+- removed "source" attribute from VG checkWindow
+- updated to @applitools/eyes-sdk-core@10.2.0
+- updated to @applitools/visual-grid-client@14.1.0
+- updated to @applitools/dom-utils@4.7.17
+
+## 4.36.1 - 2020/5/17
+
+- fetching resources with "referer" header.
+- updated to @applitools/eyes-sdk-core@10.1.2
+- updated to @applitools/visual-grid-client@14.0.0
+
 ## 4.36.0 - 2020/5/12
 
 - added devices to device emulation