+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -3,6 +3,28 @@
 ## Unreleased
 
 
+## 2.17.0 - 2020/6/9
+
+- added mobile web API support for VG
+- updated to @applitools/eyes-sdk-core@11.0.5 (from 11.0.4)
+- updated to @applitools/visual-grid-client@14.4.4 (from 14.4.3)
+
+## 2.16.0 - 2020/6/2
+
+- Unified core
+- updated to @applitools/eyes-sdk-core@11.0.2 (from v10.3.1)
+- updated to @applitools/visual-grid-client@14.4.1 (from v14.2.1)
+
+## 2.15.0 - 2020/5/19
+
+- Support accessibility validation
+- removed "source" attribute from VG checkWindow
+- updated to @applitools/eyes-sdk-core@10.1.2
+- updated to @applitools/visual-grid-client@14.0.1
+- updated to @applitools/dom-utils@4.7.17
+- updated to @applitools/eyes-sdk-core@10.2.0
+- updated to @applitools/visual-grid-client@14.1.0
+
 ## 2.14.1 - 2020/5/12
 
 - added devices to device emulation